backend/requirements.txt
asgiref==3.5.2
beautifultable==0.8.0
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-multiselectfield==0.1.12
django-solo==2.0.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
fandogh-cli==1.26.10
idna==3.4
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
tzlocal==2.1
urllib3==1.26.12
webencodings==0.5.1
websocket-client==1.4.1
zipp==3.8.1
